Fix for Xcode errors related to boost-for-react-native dependency (#2)
* Fix for Xcode errors related to boost-for-react-native dependency

Fixes the XCode errors that seem to be coming from "boost-for-react-native" project. Modified the header search path, as recommended in SO: https://stackoverflow.com/a/51794706

* Fix for xcode compilation errors due to boost-for-react-native dependency

Fixes the XCode errors that seem to be coming from "boost-for-react-native" project. Modified the header search path for debug build, as recommended in SO: https://stackoverflow.com/a/51794706

Closes #1